Api entity framework + Identity User

+++ wms-praktyki-yosi-api/wms-praktyki-yosi-api/Controllers/ProductsContoller.cs
@@ -1,4 +1,7 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Net;
+using System.Security.Claims;
 using wms_praktyki_yosi_api.Enitities;
 using wms_praktyki_yosi_api.Models;
 using wms_praktyki_yosi_api.Services;
@@ -6,12 +9,16 @@ using wms_praktyki_yosi_api.Services;
 namespace wms_praktyki_yosi_api.Controllers
 {
     [Route("api/products")]
+    [Authorize]
     public class ProductsController: ControllerBase
     {
 
         public readonly IProductService _productService;
-        public ProductsController(IProductService productService) {
+        private readonly IAccountService _accountService;
+
+        public ProductsController(IProductService productService, IAccountService accountService) {
             _productService = productService;
+            _accountService = accountService;
         }
 
         [HttpGet]
@@ -22,45 +29,79 @@ namespace wms_praktyki_yosi_api.Controllers
         }
 
         [HttpGet("{id}")]
-        public ActionResult<IEnumerable<Product>> Get([FromRoute]int id)
+        public async Task<ActionResult<Product>> Get([FromRoute]int id)
         {
+            if (!await UserIsAuthorized(User))
+                return Unauthorized();
+
             var product = _productService.GetById(id);
             if (product == null)
             {
-                return NotFound("Nie znaleziono");
+                return NotFound(1);
             }
             return Ok(product);
         }
         [HttpPut("{id}")]
-        public ActionResult<IEnumerable<Product>> Update([FromRoute] int id, [FromBody]ProductDto dto)
+        [Authorize(Roles = "Admin,Moderator")]
+        public async Task<ActionResult> Update([FromRoute] int id, [FromBody]ProductDto dto)
         {
-            var product = _productService.UpdateProduct(id,dto);
-            if (product == null)
-            {
-                return NotFound("Nie znaleziono");
-            }
-            return Ok(product);
-        }
+            if (!await UserIsAuthorized(User))
+                return Unauthorized();
 
-        [HttpDelete("{id}")]
-        public ActionResult DeleteProduct([FromRoute]int id)
-        {
-            var delete = _productService.RemoveProduct(id);
-            if (!delete)
-            {
-                return BadRequest("Niepoprawne id");
-            }
-            return Ok("pomyœlnie usuniêto");
-        }
-        [HttpPost]
-        public ActionResult AddProduct([FromBody] ProductDto dto)
-        {
-            if(!ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
-            _productService.AddNewProduct(dto);
-            return Ok("dodano pomyœlnie");
+
+
+            var IsUpdated  = _productService.UpdateProduct(id,dto);
+            if (!IsUpdated) { return NotFound(1); }
+            return Ok();
+        }
+
+        [HttpDelete("{id}")]
+        [Authorize(Roles = "Admin,Moderator")]
+        public async Task<ActionResult> DeleteProduct([FromRoute]int id)
+        {
+            if (!await UserIsAuthorized(User))
+                return Unauthorized();
+
+            var delete = _productService.RemoveProduct(id);
+            if (!delete)
+            {
+                return BadRequest(1);
+            }
+            return Ok();
+        }
+        [HttpPost]
+        [Authorize(Roles = "Admin,Moderator")]
+        public async Task<ActionResult> AddProduct([FromBody] ProductDto dto)
+        {
+            if (!await UserIsAuthorized(User))
+                return Unauthorized();
+
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(2);
+            }
+            var productId = _productService.AddNewProduct(dto);
+            return Created("/api/products/" + productId, null);
+        }
+        private async Task<bool> UserIsAuthorized(ClaimsPrincipal user)
+        {
+            var userEmail = user.FindFirst(ClaimTypes.Name);
+            if (userEmail is null)
+                return false;
+
+            try
+            {
+                var userInfo = await _accountService.GetUserInfo(userEmail.Value);
+                return true;
+            }
+            catch
+            {
+                return false;
+            }
         }
     }
